refactor(web-ui): extract conversations CRUD from api/mod.rs
Move conversation list/stream/create/get/delete/update into a sibling
`api/conversations.rs`:

- handlers: `list_conversations`, `stream_conversations`,
  `create_conversation`, `get_conversation`, `delete_conversation`,
  `update_conversation`
- request types: `CreateConversationRequest`, `UpdateConversationRequest`,
  `StreamConversationsQuery`
- response types: `ConversationSummary`, `ConversationDetail`,
  `ToolCallSummary`, `MessageSummary`
- helper constant: `TOOL_RESULT_HISTORY_LIMIT`
- 11 handler tests, including the runtime agent-id-switch case

`sse_response` flips from private to `pub(crate)` so the new module can
reuse it. `openapi.rs` paths and the router are repointed to the new
location.

No behavioural changes; `cargo test -p assistant-web-ui --lib` reports
241 passed (same as before).
